Comments Posts Users 총 merge (#10)
* add: User Entity 생성

* update: User Entity 생성자 추가

* add: SignupRequestDto 생성. Validation 적용

* update: 회원가입 기능 구현(Bcrypt 암호화 적용)

* update: User Entity에 유저를 deactivate 하는 메서드 추가

* update: User 회원탈퇴 기능 추가

* update: User Profile 조회, 수정 기능 추가

* fix: newsfeed 명을 post로 변경

* login feature 수정중6 (isTokenValid 수정중)

* refactor: UserService.java 파일 merge

* refactor: UserService.java 파일 merge

* fix: API 수정

* RefreshToken 으로 토큰 새로발급받는 기능 구현 완료

* RefreshToken 으로 토큰 새로발급받는 기능 구현 완료

* RefreshToken1

* RefreshToken2

* RefreshToken3

* logout 수정중1

* logout 수정중2

* logout 구현 완료

* fix: 메서드명 직관적이게 수정

* fix: 일관성 맞춰서 url 수정

---------

+++ src/main/java/com/sparta/newsfeedapp/service/PostService.java
@@ -2,31 +2,36 @@ package com.sparta.newsfeedapp.service;
 
 import com.sparta.newsfeedapp.dto.postRequestDto.PostRequestDto;
 import com.sparta.newsfeedapp.dto.postResponseDto.PostResponseDto;
+import com.sparta.newsfeedapp.entity.Comment;
 import com.sparta.newsfeedapp.entity.Post;
+import com.sparta.newsfeedapp.entity.User;
+import com.sparta.newsfeedapp.repository.CommentRepository;
 import com.sparta.newsfeedapp.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
+import org.hibernate.type.descriptor.java.ObjectJavaType;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
 public class PostService {
 
     public final PostRepository postRepository;
+    public final CommentRepository commentRepository;
 
-    public PostResponseDto createPost(PostRequestDto requestDto) {
+    public PostResponseDto createPost(PostRequestDto requestDto, User user) {
         // 인가 기능 추가하기
-
-        Post post = new Post(requestDto);
+        Post post = new Post(requestDto, user);
 
         postRepository.save(post);
 
         return new PostResponseDto(post);
     }
 
-    public List<PostResponseDto> getPostAll() {
+    public List<PostResponseDto> getAllPost() {
         return postRepository.findAllByOrderByCreatedAtDesc().stream().map(PostResponseDto :: new).toList();
     }
 
@@ -34,25 +39,26 @@ public class PostService {
         return postRepository.findAllById(id).stream().map(PostResponseDto :: new).toList();
     }
 
+    public List<Comment> getComments(Long postId){
+        return commentRepository.findByPostId(postId);
+    }
+
     @Transactional
-    public PostResponseDto updatePost(Long id, PostRequestDto requestDto) {
+    public PostResponseDto updatePost(Long id, PostRequestDto requestDto, User user) {
         Post post = findPostById(id);
-
-        // 인가 기능 추가하기
-
+        if (!Objects.equals(post.getUser(), user)){
+            throw new IllegalArgumentException("본인 게시글만 수정할 수 있습니다.");
+        }
         post.update(requestDto);
-
         return new PostResponseDto(post);
     }
 
-    public Long deletePost(Long id) {
+    public Long deletePost(Long id, User user) {
         Post post = findPostById(id);
-
-        // 인가 기능 추가하기
-
-
+        if (!Objects.equals(post.getUser(), user)){
+            throw new IllegalArgumentException("본인 게시글만 수정할 수 있습니다.");
+        }
         postRepository.delete(post);
-
         return id;
     }
